amesie_ai_backend/requirements-light.txt
# Lightweight requirements for initial deployment (without ML models)
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.8.2
python-multipart==0.0.6
python-dotenv==1.0.0
psutil==5.9.6
websockets==12.0
httpx>=0.27.0
prometheus-client==0.19.0
structlog==23.2.0
gunicorn==21.2.0